requirements.txt
# Core dependencies
tradingview-scraper==0.4.19
pandas>=2.0.3
numpy>=1.24.0
yfinance>=0.2.28
ta>=0.11.0



# Google Sheets integration
gspread>=5.12.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1

# Configuration
PyYAML>=6.0.1

# Rate limiting and async
ratelimit>=2.2.1

# Testing
pytest>=7.4.3
pytest-mock>=3.12.0

# Data validation
pydantic>=2.8.2

# Logging
colorlog>=6.8.0

# Progress tracking
tqdm>=4.66.1
